// src/espIdf/unitTest/types.ts
/*
 * Project: ESP-IDF VSCode Extension
 * File Created: Thursday, 6th July 2023 7:36:13 pm
 * Copyright 2023 Espressif Systems (Shanghai) CO LTD
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { TestItem, Uri } from "vscode";

export interface EspIdfTestItem {
  id: string;
  testName: string;
  type: string;
  label: string;
  children?: EspIdfTestItem[];
  tooltip?: string;
  filePath?: string;
  uri?: Uri;
}

export const idfTestData = new WeakMap<TestItem, EspIdfTestItem>();